requirements.txt
# Core dependencies
openai>=1.0.0
requests>=2.31.0
pyyaml>=6.0

# Utilities
python-dotenv>=1.0.0

# Memory/Embeddings (Milestone 2)
sentence-transformers>=2.2.0
chromadb>=0.4.0

# Voice (Milestone 3)
# Note: OpenAI TTS/STT uses openai package above

# Image (Milestone 4)
pillow>=10.0.0

# API Server (Milestone 5)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Web UI (Milestone 5)
streamlit>=1.28.0
